Modified Code to implement device toggling
I modified the codebase to more accuratly communicate with the Arduino for the purpose of toggling devices. This included modifications to the configuration file for the pin locations of the devices. This is to prevent the need to do string matching on the Arduino. This reduces the size of the codebase for the Arduino. It also allows for the pinout of the devices to be changed 'on the fly' if that becomes a requirement of the project at a later date.